+++ src/Core/GUI/SwingEX/EXColorDefaults.java
@@ -13,5 +13,8 @@ public interface EXColorDefaults {
     Color NONE = new Color(0,0,0,0);
     Color BLACK = new Color(0,0,0,255);
     Color NULL = new Color(200, 80, 255, 255);
+    Color WHITE = new Color(255, 255, 255, 255);
+    Color GREY = new Color(120, 120, 120, 255);
+    Color CRIMSON = new Color(130, 30, 30, 255);
 
 }