+++ src/Core/GUI/SwingEX/XConstants.java
@@ -10,13 +10,7 @@ public interface XConstants {
 
     int HORIZONTAL_SORT = 1;
     int VERTICAL_SORT = 2;
-    int ORIENT_LEFT = 3;
-    int ORIENT_RIGHT = 4;
-    int ORIENT_TOP = 5;
-    int ORIENT_BOTTOM = 6;
-    int ORIENT_CENTER = 7;
-    int VERTICAL_SORT_REVERSE = 8;
-    int HORIZONTAL_SORT_REVERSE = 9;
-
+    int VERTICAL_SORT_REVERSE = 3;
+    int HORIZONTAL_SORT_REVERSE = 4;
 
 }