+++ src/Core/GUI/gfxConstants.java
@@ -1,0 +1,30 @@
+package Core.GUI;
+
+import java.awt.*;
+
+/**
+ * KM
+ * June 22 2017
+ * Handles the graphic constants.
+ */
+
+public interface gfxConstants {
+
+    Color clrText = new Color(255, 255, 255, 255);
+    Color clrTextTranslucent = new Color(255, 255, 255, 150);
+    Color clrInvisible = new Color(0, 0, 0, 0);
+    Color clrTrueBlack = new Color(0, 0, 0, 255);
+    Color clrBlkTransparent = new Color(15, 35, 25, 175);
+    Color clrBlk = new Color(25, 35, 35, 255);
+    Color clrDGrey = new Color(47, 80, 68, 255);
+    Color clrDisableBorder = new Color(75, 5, 25, 255);
+    Color clrDisable = new Color(135, 15, 55, 255);
+    Color clrEnable = new Color(108, 255, 224, 255);
+    Color clrDark = new Color(70, 170, 160, 255);
+    Color clrButtonBackground = new Color(0, 125, 90, 255);
+    Color clrBGOpaque = new Color(25, 90, 60, 255);
+    Color clrButtonMain = new Color(0, 155, 90, 255);
+    Color clrBackground = new Color(54, 95, 77, 105);
+    Color clrForeground = new Color(0, 185, 110, 155);
+
+}